+++ fpdf/outline.py
@@ -110,6 +110,14 @@ def build_outline_objs(sections):
 
 
 class TableOfContents:
+    """
+    A reference implementation of a Table of Contents (ToC) for use with `fpdf2`.
+
+    This class provides a customizable Table of Contents that can be used directly or subclassed
+    for additional functionality. To use this class, create an instance of `TableOfContents`, 
+    configure it as needed, and pass its `render_toc` method as the `render_toc_function` argument 
+    to `FPDF.insert_toc_placeholder()`.
+    """
 
     def __init__(self):
         self.text_style = TextStyle()